Move settings controller to admin
