+++ app/controllers/admin/settings_controller.rb
@@ -1,0 +1,12 @@
+module Admin
+  class SettingsController < ApplicationController
+    before_action :authorize_admin
+
+    def index
+      @maximum_diagnoses        = Setting.maximum_diagnoses
+      @maximum_prescriptions    = Setting.maximum_prescriptions
+      @medical_history_sequence = Setting.medical_history_sequence
+      @emergency_number         = Setting.emergency_number
+    end
+  end
+end
