+++ app/views/admin/settings/index.haml
@@ -1,5 +1,5 @@
 = content_for :nav_option do
-  = active_nav_to('Preferencias', settings_path)
+  = active_nav_to('Preferencias', admin_settings_path)
 
 .row
   .col-md-12
