+++ app/views/layouts/application.haml
@@ -42,7 +42,7 @@
                   %ul.dropdown-menu{ role: "menu" }
                     = nav_to 'Actividades', admin_activities_path
                     = nav_to 'Cobros', admin_charges_path
-                    = nav_to 'Preferencias', settings_path
+                    = nav_to 'Preferencias', admin_settings_path
                     = nav_to 'Sucrusales', admin_branch_offices_path
                     = nav_to 'Usuarios', admin_users_path
               = yield :nav_option
